+++ packages/maimai-chart-engine/src/core/index.ts
@@ -1,6 +1,8 @@
 export {
   parseSimaiChart,
+  parseMa2Chart,
   getAvailableDifficulties,
+  type ChartFileType,
   default as ChartParser,
 } from "./parser/ChartParser";
 export { AudioManager, type AudioManagerConfig } from "./audio/AudioManager";
